Move settings to separate file.
